requirements.txt
python-telegram-bot==20.6
openai>=1.0.0
python-dotenv==1.0.0
pytz==2023.3
aiohttp==3.8.6 